pyproject.toml
[tool.poetry]
name = "recipts"
version = "0.1.0"
description = ""
authors = ["diogoazevedo15 <68329635+diogoazevedo15@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7"
streamlit = "^1.40.1"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.1"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.154.0"
google-cloud-vision = "^3.8.1"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
